removing postId from user colletions when deleted

+++ controllers/feed.js
@@ -126,6 +126,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if(post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized");
+        error.statusCode = 403;
+        throw error;
+      } 
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -157,12 +162,22 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-      // Check logged in user
+      if(post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized");
+        error.statusCode = 403;
+        throw error;
+      } 
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
     })
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: "Deleted post." });
     })
     .catch((err) => {
